fix(server-settings): move auto-start options to "Settings"
